src/edge/EdgeSolution/modules/kanai/requirements-openvino.txt
FastAPI
python-multipart
opencv-python
uvicorn
numpy
pillow
pydantic
networkx
requests
httpx
azure-iot-device
pytz
grpcio
protobuf==3.19.*